requirements/dev.txt
-r base.txt
ipython>=8.0
django-debug-toolbar>=4.0
